pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meowtv"
version = "1.0.0"
description = "CLI for streaming content from MeowTV providers"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "MeowTV" }
]
keywords = ["streaming", "cli", "meowtv"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pycryptodome>=3.20.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "yt-dlp>=2024.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
meowtv = "meowtv.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["meowtv"]
